Add redirects
Setups up a new global field list where you can add redirects which
take a url pattern and will send the request to the "to" pattern. This
is placed in global site data as specified in issue as these are global
slugs to redirect.

I decided to keep this in the app file as it seemed the best global
place to add logic to every page on the site. The redirect functionality
was more or less copied from the old forestry.io site.

Fixes # 20.